code/requirements.txt
google-generativeai>=0.7.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
rich>=13.0.0
